Working simulation (!)
Working flight altitude simulation using Stan, with all the bells and whistles (including estimation of state probabilities). See stan_attempt_7. Don't expect any miracles here- not a lot of data, so I'll have to be a little gentle with it. That being said, not bad at all for a first Stan model.