Mu/Tau parameterization, beta distribution prior
Switched over to the mu/tau parameterization of the gamma distribution, which increased neff and allowed for an informative prior on the mean flight altitude. Also installed a vague prior of beta(2,2) for the proportion of flight locations, which should decrease the likelihood of extreme values. Proportion of flight locations down to 0.74, and mean flight altitude up to 305 m. Still seems a little high/low, respectively. Going to try some simulations to see if this model can retrieve this parameter effectively.