Attempt at a cauchy distribution
Seeing if a cauchy distribution is a better fit for the observation error than the normal distribution. Short answer- not really. The shape of the distribution is a little better, but it's not as peaky as it should be, and all the test statistics are way off. Gonna stick with a normal distribution from here on out.